seperate text && sentence query for lemma attestation counts
